requirements.txt
streamlit==1.38.0
google-generativeai==0.8.3
qdrant-client==1.12.0
sentence-transformers
python-dotenv==1.0.1
sqlite-utils==3.37
scrapy==2.11.1
trafilatura==1.12.2
jaydebeapi==1.2.3
slack-sdk==3.27.1
atlassian-python-api==3.41.3
langchain
langchain-google-genai
Numpy
